+++ libs/webb-ui-components/src/components/Link/index.ts
@@ -1,0 +1,2 @@
+export * from './Link';
+export * from './types';
