--- libs/webb-ui-components/src/components/SideBar/components/DropdownControl.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { ChevronDown, ChevronUp } from '@webb-tools/icons';
-
-// Types
-type DropdownControlProps = {
-  isDropdownOpen: boolean;
-};
-
-// Component
-export const DropdownControl: React.FC<DropdownControlProps> = ({
-  isDropdownOpen,
-}) => {
-  return <div>{isDropdownOpen ? <ChevronDown /> : <ChevronUp />}</div>;
-};
