+++ libs/webb-ui-components/src/components/SideBar/index.ts
@@ -1,1 +1,2 @@
 export * from './SideBar';
+export * from './types';
