--- libs/webb-ui-components/src/components/SideBar/utils.ts
@@ -1,15 +1,0 @@
-export const pushToInternalLink = (href: string) => {
-  const isHashRouting = window.location.hash.length > 0;
-
-  if (isHashRouting) {
-    // React
-    window.location.hash = '#' + href;
-  } else {
-    // NextJS
-    window.location.pathname = href;
-  }
-};
-
-export const pushToExternalLink = (href: string) => {
-  window.open(href, '_blank');
-};
